Add working server application code
Add logic for post functionality in server and add also test cases for the function.

Add target in make file for building the software, cleaning the directories
and to run the server application test cases.

Add configuration parameters in vm.args file which will be avaiable as
environment variables in cloud backup server application so that the
hard coding of ip address, tcp port number can be avoided.